feat: construct evidence set by clue set(one attribute threshold one bit)

+++ src/main/java/ddfinder/predicate/PredicateProvider.java
@@ -10,14 +10,14 @@ import java.util.Map;
  * @author tristonK 2022/12/31
  */
 public class PredicateProvider {
-    private final Map<Operator, Map<ColumnOperand<?>, Map<Integer, Predicate>>> predicates;
+    private final Map<Operator, Map<ColumnOperand<?>, Map<Double, Predicate>>> predicates;
 
     public PredicateProvider() {
         predicates = new HashMap<>();
     }
 
-    public Predicate getPredicate(Operator op, ColumnOperand<?> op1, Integer distance) {
-        Map<Integer, Predicate> map = predicates.computeIfAbsent(op, a -> new HashMap<>()).computeIfAbsent(op1, a -> new HashMap<>());
+    public Predicate getPredicate(Operator op, ColumnOperand<?> op1, Double distance) {
+        Map<Double, Predicate> map = predicates.computeIfAbsent(op, a -> new HashMap<>()).computeIfAbsent(op1, a -> new HashMap<>());
         Predicate p = map.get(distance);
         if (p == null) {
             p = new Predicate(op, distance, op1);
